+++ batch/batch.js
@@ -3,10 +3,11 @@
 var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 var debug = require('./util/debug')('batch');
-var forever = require('./util/forever');
 var queue = require('queue-async');
-var Locker = require('./leader/locker');
 var HostUserQueueMover = require('./maintenance/host-user-queue-mover');
+var HostScheduler = require('./scheduler/host-scheduler');
+
+var EMPTY_QUEUE = true;
 
 function Batch(name, jobSubscriber, jobQueue, jobRunner, jobService, jobPublisher, redisPool, logger) {
     EventEmitter.call(this);
@@ -17,11 +18,11 @@ function Batch(name, jobSubscriber, jobQueue, jobRunner, jobService, jobPublishe
     this.jobService = jobService;
     this.jobPublisher = jobPublisher;
     this.logger = logger;
-    this.locker = Locker.create('redis-distlock', { pool: redisPool });
+    this.hostScheduler = new HostScheduler(name, { run: this.processJob.bind(this) }, redisPool);
     this.hostUserQueueMover = new HostUserQueueMover(jobQueue, jobService, this.locker, redisPool);
 
-    // map: host => jobId
-    this.workingQueues = {};
+    // map: user => jobId. Will be used for draining jobs.
+    this.workInProgressJobs = {};
 }
 util.inherits(Batch, EventEmitter);
 
@@ -38,37 +39,15 @@ Batch.prototype.subscribe = function () {
 
     this.jobSubscriber.subscribe(
         function onJobHandler(user, host) {
-            var resource = host + ':' + user;
-            debug('onJobHandler(%s, %s)', user, host);
-            if (self.isProcessingUser(user)) {
-                return debug('%s is already processing user=%s', self.name, user);
-            }
-
-            // do forever, it does not throw a stack overflow
-            forever(
-                function (next) {
-                    self.locker.lock(resource, function(err) {
-                        // we didn't get the lock for the host
-                        if (err) {
-                            debug(
-                                'Could not lock host=%s for user=%s from %s. Reason: %s',
-                                host, self.name, user, err.message
-                            );
-                            return next(err);
-                        }
-                        debug('Locked host=%s for user=%s from %s', host, user, self.name);
-                        self.processNextJob(user, next);
-                    });
-                },
-                function (err) {
-                    if (err) {
-                        debug(err.name === 'EmptyQueue' ? err.message : err);
-                    }
-
-                    self.finishedProcessingUser(user);
-                    self.locker.unlock(resource, debug);
+            debug('[%s] onJobHandler(%s, %s)', self.name, user, host);
+            self.hostScheduler.add(host, user, function(err) {
+                if (err) {
+                    return debug(
+                        'Could not schedule host=%s user=%s from %s. Reason: %s',
+                        host, self.name, user, err.message
+                    );
                 }
-            );
+            });
         },
         function onJobSubscriberReady(err) {
             if (err) {
@@ -80,50 +59,45 @@ Batch.prototype.subscribe = function () {
     );
 };
 
-Batch.prototype.processNextJob = function (user, callback) {
-    // This is missing the logic for processing several users within the same host
-    // It requires to:
-    //  - Take care of number of jobs running at the same time per host.
-    //  - Execute user jobs in order.
+Batch.prototype.processJob = function (user, callback) {
     var self = this;
     self.jobQueue.dequeue(user, function (err, jobId) {
         if (err) {
-            return callback(err);
+            return callback(new Error('Could not get job from "' + user + '". Reason: ' + err.message), !EMPTY_QUEUE);
         }
 
         if (!jobId) {
-            var emptyQueueError = new Error('Queue for user="' + user + '" is empty');
-            emptyQueueError.name = 'EmptyQueue';
-            return callback(emptyQueueError);
+            debug('Queue empty user=%s', user);
+            return callback(null, EMPTY_QUEUE);
         }
 
-        self.setProcessingJobId(user, jobId);
-
+        self.setWorkInProgressJob(user, jobId);
         self.jobRunner.run(jobId, function (err, job) {
-            self.setProcessingJobId(user, null);
+            self.clearWorkInProgressJob(user);
 
             if (err) {
                 debug(err);
                 if (err.name === 'JobNotRunnable') {
-                    return callback();
+                    return callback(null, !EMPTY_QUEUE);
                 }
-                return callback(err);
+                return callback(err, !EMPTY_QUEUE);
             }
 
-            debug('Job=%s status=%s user=%s (failed_reason=%s)', jobId, job.data.status, user, job.failed_reason);
+            debug(
+                '[%s] Job=%s status=%s user=%s (failed_reason=%s)',
+                self.name, jobId, job.data.status, user, job.failed_reason
+            );
 
             self.logger.log(job);
 
-            self.emit('job:' + job.data.status, jobId);
-
-            callback();
+            return callback(null, !EMPTY_QUEUE);
         });
     });
 };
 
 Batch.prototype.drain = function (callback) {
     var self = this;
-    var workingUsers = this.getWorkingUsers();
+    var workingUsers = this.getWorkInProgressUsers();
     var batchQueues = queue(workingUsers.length);
 
     workingUsers.forEach(function (user) {
@@ -143,7 +117,7 @@ Batch.prototype.drain = function (callback) {
 
 Batch.prototype._drainJob = function (user, callback) {
     var self = this;
-    var job_id = this.getProcessingJobId(user);
+    var job_id = this.getWorkInProgressJob(user);
 
     if (!job_id) {
         return process.nextTick(function () {
@@ -169,22 +143,21 @@ Batch.prototype.stop = function (callback) {
     this.jobSubscriber.unsubscribe(callback);
 };
 
-Batch.prototype.isProcessingUser = function(user) {
-    return this.workingQueues.hasOwnProperty(user);
+
+/* Work in progress jobs */
+
+Batch.prototype.setWorkInProgressJob = function(user, jobId) {
+    this.workInProgressJobs[user] = jobId;
 };
 
-Batch.prototype.getWorkingUsers = function() {
-    return Object.keys(this.workingQueues);
+Batch.prototype.getWorkInProgressJob = function(user) {
+    return this.workInProgressJobs[user];
 };
 
-Batch.prototype.setProcessingJobId = function(user, jobId) {
-    this.workingQueues[user] = jobId;
+Batch.prototype.clearWorkInProgressJob = function(user) {
+    delete this.workInProgressJobs[user];
 };
 
-Batch.prototype.getProcessingJobId = function(user) {
-    return this.workingQueues[user];
-};
-
-Batch.prototype.finishedProcessingUser = function(user) {
-    delete this.workingQueues[user];
+Batch.prototype.getWorkInProgressUsers = function() {
+    return Object.keys(this.workInProgressJobs);
 };